Make the terminating sort unconditional rather than a downlevel polyfill
Polyfills.Order backfilled a merge sort for the frameworks that lack
Enumerable.Order, and the claim that came with it was that
Array.prototype.sort, toSorted and %TypedArray%.prototype.sort behave
identically on all five target frameworks. They do not. The guard was
#if !NET8_0_OR_GREATER, so net8.0 and net10.0 still bind Enumerable.Order and
still sort with the BCL introsort.

That introsort does not spin the way .NET Framework's quicksort does, but it
does not finish either: it detects the inconsistency and throws
ArgumentException. An inconsistent comparison function is legal JavaScript --
https://tc39.es/ecma262/#sec-sortcompare leaves the resulting order
implementation-defined and https://tc39.es/ecma262/#sec-sortindexedproperties
still requires the sort to complete. And ArgumentException is not a
JavaScriptException, while the three call sites unwrap only
InvalidOperationException, so `try { a.sort(function () { return -1; }); }
catch (e) {}` never caught it and the exception escaped engine.Evaluate
entirely.

The merge sort therefore leaves Polyfills, where an unconditional helper does
not belong, and becomes SortExtensions.StableOrder. Being the one
implementation on every target framework is also what finally makes the
identical-order claim true. It hands back the array it sorted instead of an
IEnumerable, so the two call sites that wanted an array stop copying it a
second time.

The regression test passed vacuously above net462: 20 elements is over
.NET Core's 16-element insertion-sort threshold, but `a === 1 ? -1 : 1`
happens never to trip the detector. It now also runs the blunter `return -1`,
which failed all three sorts on net10.0 before this change and none after.
